% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quantile_add}
\alias{quantile_add}
\title{Add an image to the stack from which to compute quantiles.}
\usage{
quantile_add(base, image)
}
\arguments{
\item{base}{The object that `build_tierney` made.}

\item{image}{A two-dimensional array of numeric. Can include NaN values.}
}
\value{
A modified base object, that you pass in the next time.
}
\description{
Add an image to the stack from which to compute quantiles.
}
