% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tierney.R
\name{quantiles}
\alias{quantiles}
\title{Retrieve median and confidence intervals.}
\usage{
quantiles(base)
}
\arguments{
\item{base}{A Tierney quantile estimator.}
}
\description{
Retrieve median and confidence intervals.
}
