% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tierney.R
\name{build_tierney}
\alias{build_tierney}
\title{Create an object to gather quantiles.}
\usage{
build_tierney(extent, total, level = 95)
}
\arguments{
\item{extent}{The dimensions of the image as c(rows, cols).}

\item{total}{The total number of images from which to calculate quantiles.}

\item{level}{The confidence interval, e.g. 95 or 90.}
}
\description{
Create an object to gather quantiles.
}
