% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tierney.R
\name{quantiles}
\alias{quantiles}
\title{Retrieve median and confidence intervals.}
\usage{
quantiles(base)
}
\arguments{
\item{base}{A Tierney quantile estimator.

If every image had a NaN for a pixel, that pixel's quantiles will be all NaN.
If the pixel had some NaN values, the quantiles will be adjusted to account
for those fewer values.}
}
\description{
Retrieve median and confidence intervals.
}
